requirements.txt
flask==2.3.3
flask-cors==4.0.0
bcrypt
pyotp
python-dotenv==1.0.0
flask-jwt-extended==4.6.0
Werkzeug==2.3.7

